Fix asset source path for craft instances in sub directories
If craft runs in a subdirectory imager creates this subdirectory also as a folder.
The consequence is that on different environments this could lead to multiple transformed folders.

`http://localhost/craftWebsite`
this will create: `public/imager/craftWebsite/img...`

`http://localhost/otherEnv`
-> `public/imager/otherEnv/img...`

This is an issue for deployments of assets. So this part will be removed and imager should create the files relatively to the assets path.